refactor: socketio
