+++ .gitignore
@@ -16,3 +16,4 @@ Cargo.lock
 
 .vscode
 .idea
+.env
