+++ compose.dev.yaml
@@ -7,5 +7,5 @@ services:
     restart: always
     network_mode: host
     user: root
-    environment:
-      - NCA_API_KEY=${NCA_API_KEY}
+    env_file:
+      - .env
