+++ src/cli.rs
@@ -3,8 +3,8 @@ use clap::Parser;
 #[derive(Clone, Parser)]
 #[command(
     name = "NodeCookAgent",
-    version = "0.1.0",
-    author = "long2ice",
+    version = env!("CARGO_PKG_VERSION"),
+    author = "NodeCook <dev@nodecook.com>",
     about = "Agent for NodeCook to run jobs"
 )]
 pub struct Cli {
@@ -20,16 +20,10 @@ pub struct Cli {
     /// IPv6 only mode
     #[arg(long, default_value_t = false, env = "NCA_IPV6_ONLY")]
     pub ipv6_only: bool,
-    /// Port to listen on
-    #[arg(short, long, default_value_t = 4000, env = "NCA_PORT")]
-    pub port: u16,
     /// API key comes from nodecook to know this node belongs to you
     #[arg(short, long, env = "NCA_API_KEY")]
     pub api_key: String,
     /// Enable debug mode
     #[arg(short, long, default_value_t = false, env = "NCA_DEBUG")]
     pub debug: bool,
-    /// Endpoint for agent to access, default is host ip:port, if you are behind proxy, you should set this to your public address
-    #[arg(short, long, env = "NCA_ENDPOINT")]
-    pub endpoint: Option<String>,
 }
