+++ src/handlers/mod.rs
@@ -1,0 +1,5 @@
+pub mod dns;
+pub mod http;
+pub mod mtr;
+pub mod ping;
+pub mod tcping;
