+++ src/main.rs
@@ -1,67 +1,59 @@
-mod api;
-mod app;
 mod cli;
+mod client;
 mod constant;
-mod dns;
+mod dns_resolve;
 mod errors;
 mod handlers;
 mod utils;
-use crate::app::create_app;
-use crate::cli::Cli;
-use crate::utils::add_agent_with_args;
+use crate::constant::V4_SERVER;
+use crate::{cli::Cli, constant::V6_SERVER};
+
+use crate::client::connect_server;
 use clap::Parser;
 use tokio::signal;
-use tokio_cron_scheduler::{Job, JobScheduler};
-use tracing::{info, Level};
+use tracing::{debug, error, info, Level};
 use tracing_subscriber::fmt;
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
+async fn main() {
     let mut level = Level::INFO;
     let args = Cli::parse();
-    let api_key = args.api_key.clone();
-    let port = args.port;
     if args.debug {
         level = Level::DEBUG;
     }
     let collector = fmt().with_max_level(level).finish();
-    tracing::subscriber::set_global_default(collector)?;
+    tracing::subscriber::set_global_default(collector).unwrap();
     if args.ipv4_only && args.ipv6_only {
-        panic!("ipv4_only and ipv6_only can't be true at the same time");
+        error!("ipv4_only and ipv6_only can't be true at the same time");
+        std::process::exit(1);
     }
     let mut v4_ok = false;
     let mut v6_ok = false;
     if !args.ipv4_only {
-        v6_ok = add_agent_with_args(args.clone(), "ipv6", true).await;
+        let server = args.ipv6_server.unwrap_or_else(|| V6_SERVER.to_string());
+        let ret = connect_server(server, args.api_key.clone()).await;
+        if ret.is_ok() {
+            v4_ok = true
+        } else {
+            debug!("connect ipv6 server failed: {}", ret.err().unwrap());
+        }
     }
     if !args.ipv6_only {
-        v4_ok = add_agent_with_args(args.clone(), "ipv4", true).await;
+        let server = args.ipv4_server.unwrap_or_else(|| V4_SERVER.to_string());
+        let ret = connect_server(server, args.api_key).await;
+        if ret.is_ok() {
+            v6_ok = true
+        } else {
+            debug!("connect ipv4 server failed: {}", ret.err().unwrap());
+        }
     }
     if !v4_ok && !v6_ok {
-        panic!("add ipv4 and ipv6 agent failed, please check your network or try again later");
+        error!("connect ipv4 and ipv6 server failed, please check your network or try again later");
+        std::process::exit(1);
     }
-    let sched = JobScheduler::new().await?;
-    sched
-        .add(Job::new_async("*/60 * * * * *", move |_uuid, _l| {
-            let args = args.clone();
-            Box::pin(async move {
-                if v4_ok && !args.ipv6_only {
-                    add_agent_with_args(args.clone(), "ipv4", false).await;
-                }
-                if v6_ok && !args.ipv4_only {
-                    add_agent_with_args(args, "ipv6", false).await;
-                }
-            })
-        })?)
-        .await?;
-    sched.start().await?;
-    let addr = format!(":::{}", port);
-    info!("listening on addr {}", addr);
-    let listener = tokio::net::TcpListener::bind(addr).await?;
-    axum::serve(listener, create_app(api_key))
-        .with_graceful_shutdown(shutdown_signal())
-        .await?;
-    Ok(())
+    info!("Congratulations, you have successfully started the agent!");
+
+    shutdown_signal().await;
 }
 
 async fn shutdown_signal() {
@@ -81,7 +73,6 @@ async fn shutdown_signal() {
 
     #[cfg(not(unix))]
     let terminate = std::future::pending::<()>();
-
     tokio::select! {
         _ = ctrl_c => {},
         _ = terminate => {},
